Harden dashboard auth follow-up coverage
