+++ packages/node-ui/package.json
@@ -12,6 +12,7 @@
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:e2e": "playwright test",
+    "test:e2e:static": "playwright test --config playwright.static.config.ts",
     "test:e2e:hw-mock": "playwright test --config playwright.hw-mock.config.ts",
     "test:e2e:headed": "playwright test --headed",
     "test:e2e:ui": "playwright test --ui",
